requirements.txt
asgiref==3.8.1
Django==5.1.4
django-filter==24.3
django-storages==1.14.4
mysqlclient==2.2.6
pillow==11.1.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2024.2
